[PP-10] Add main page and admin route
