+++ frontend/package.json
@@ -14,6 +14,7 @@
     "material-react-table": "^0.15.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
