+++ frontend/src/App.js
@@ -5,59 +5,15 @@ import React, { useState, useEffect } from "react";
 import { useMemo } from 'react';
 import MaterialReactTable from 'material-react-table';
 import Sidebar from './components/Sidebar'
+import { Outlet, Link } from "react-router-dom";
 
 function App() {
-  const [remoteData, setRemoteData] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true);
-      const response = await fetch(
-        'http://localhost:8080/api/category',
-      );
-      const json = await response.json();
-      setRemoteData(json);
-      setIsLoading(false);
-    };
-    fetchData();
-  }, []);
-
-  const parsedData = useMemo(
-    () =>
-      remoteData.map((userData) => ({
-        categoryId: userData.categoryId,
-        categoryName: userData.categoryName,
-      })) ?? [],
-    [remoteData],
-  );
-
-  const columns = useMemo(
-    () => [
-      {
-        header: 'ID',
-        id: 'categoryId',
-      },
-      {
-        header: 'Name',
-        id: 'categoryName',
-      },
   
-    ],
-    [],
-  );
 
   return (
     <div className="main">
-    <Sidebar />
+        <Link to="/admin">Invoices</Link>
     <div className="container">
-    <MaterialReactTable
-      columns={columns}
-      data={parsedData}
-      enableDensityToggle
-      state={{
-        isLoading,
-      }} />
           </div>
     </div>
   );
